chore: moved helpers into it's own file; repaired imports/refs
